Add files via upload
This is the obtained result. It is a graph that consists days as X-Axis and Price(in USD) as Y axis. The graph depicts the rise and fall in the prices of the stock for a certain time period measured in Days. Based on these records, we can predict the future stock prices using LSTM algorithm which plays a major role in Artificial Intelligence.